requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
dashscope==1.24.2
websocket-client==1.6.1
tenacity==8.2.3
openai==1.3.5
google-generativeai==0.3.2
pyttsx3==2.90
SpeechRecognition==3.10.0